# api/openapi.yaml
openapi: 3.1.0
info:
  title: Puerta Abierta API
  version: 1.0.0
  summary: STR operations API for Paraguay-focused owners and operators
  description: |
    REST API contract for multi-tenant short-term rental operations:
    organizations, properties, reservations, tasks, finance, and reporting.
servers:
  - url: https://api.puertaabierta.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /properties:
    get:
      tags: [Properties]
      summary: List properties
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/PropertyStatus"
      responses:
        "200":
          description: Property list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Property"
                required: [data]
    post:
      tags: [Properties]
      summary: Create property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyInput"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{property_id}:
    get:
      tags: [Properties]
      summary: Get property by id
      parameters:
        - $ref: "#/components/parameters/PropertyIdPath"
      responses:
        "200":
          description: Property detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Properties]
      summary: Update property
      parameters:
        - $ref: "#/components/parameters/PropertyIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyInput"
      responses:
        "200":
          description: Property updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /units:
    get:
      tags: [Units]
      summary: List units
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: property_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Unit list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Unit"
                required: [data]
    post:
      tags: [Units]
      summary: Create unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitInput"
      responses:
        "201":
          description: Unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{unit_id}:
    get:
      tags: [Units]
      summary: Get unit by id
      parameters:
        - $ref: "#/components/parameters/UnitIdPath"
      responses:
        "200":
          description: Unit detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Units]
      summary: Update unit
      parameters:
        - $ref: "#/components/parameters/UnitIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUnitInput"
      responses:
        "200":
          description: Unit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
...
  /reports/owner-summary:
    get:
      tags: [Reports]
      summary: Owner summary report
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: property_id
          in: query
          schema:
            type: string
            format: uuid
        - name: unit_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Summary metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerSummaryReport"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrgIdQuery:
      name: org_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
    PropertyIdPath:
      name: property_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UnitIdPath:
      name: unit_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    UUID:
      type: string
      format: uuid
    Currency:
      type: string
      enum: [PYG, USD]
...
    PropertyStatus:
      type: string
      enum: [active, inactive]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required: [error, message]
...
    Property:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        code:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/PropertyStatus"
        city:
          type: string
        country_code:
          type: string
      required: [id, organization_id, name, status, city, country_code]
    Unit:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        property_id:
          $ref: "#/components/schemas/UUID"
        code:
          type: string
        name:
          type: string
        max_guests:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: number
        currency:
          $ref: "#/components/schemas/Currency"
        is_active:
          type: boolean
      required:
        [id, organization_id, property_id, code, name, max_guests, bedrooms, bathrooms, currency, is_active]
...
    OwnerSummaryReport:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        occupancy_rate:
          type: number
          description: Value from 0 to 1
        gross_revenue:
          type: number
        expenses:
          type: number
        net_payout:
          type: number
      required: [organization_id, from, to, occupancy_rate, gross_revenue, expenses, net_payout]
...
    CreatePropertyInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        code:
          type: string
        status:
          $ref: "#/components/schemas/PropertyStatus"
        address_line1:
          type: string
        city:
          type: string
      required: [organization_id, name]
    UpdatePropertyInput:
      type: object
      properties:
        name:
          type: string
        status:
          $ref: "#/components/schemas/PropertyStatus"
        address_line1:
          type: string
        city:
          type: string
    CreateUnitInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        property_id:
          $ref: "#/components/schemas/UUID"
        code:
          type: string
        name:
          type: string
        max_guests:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: number
        currency:
          $ref: "#/components/schemas/Currency"
      required: [organization_id, property_id, code, name]
    UpdateUnitInput:
      type: object
      properties:
        name:
          type: string
        max_guests:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: number
        is_active:
          type: boolean
...